Add GPU support. (#1)
It would be cool to be able to leverage GPU to compute pitch much faster. Currently this does not work because of incorrect device placement for new tensors. However, with some simple changes the existing implementation is GPU-compatible. The speedup is 50x on my machine for a batch of 32, 10 second waves.